refactor: 공통 타입들 DomainCommon, FeatureCommon으로 책임 분리 - #278
